+++ api/annexure_details/annexure_details.service.js
@@ -3,10 +3,13 @@ var async = require('async');
 const DATE_FORMATTER=require('dateformat')
 module.exports={
     createAnnexure_details:(data,callBack)=>{
+        console.log("Getting request :-", data);
         var cur=new Date().toLocaleString('en-US',{timeZone:'Asia/Calcutta'});
         created_date=DATE_FORMATTER(cur,"yyyy-mm-dd hh:MM:ss");
         updated_date=DATE_FORMATTER(cur,"yyyy-mm-dd hh:MM:ss");
-        var prodinfo=data.productinfo;
+        var prodinfo=data.prodinfo;
+        var acinfo=data.acinfo;
+        if(data.prodinfo.length != 0 || data.acinfo.length != 0){
         pool.query(
             `insert into annexure (purchase_order_id,created_date,updated_date,status) values(?,?,?,1)`,
             [
@@ -21,38 +24,113 @@ module.exports={
                 }
                 if(results.insertId != 0){
                     var anid=results.insertId;
-                    console.log(prodinfo,"data");
+                    // console.log(data.prodinfo,"data");
                     var len=prodinfo.length;
-                    prodinfo.forEach(element => {
-                        pool.query(
-                            `insert into annexure_details (annexure_id,product_id,length,quantity,total_length,module,area,created_date,updated_date,status) values(?,?,?,?,?,?,?,?,?,1)`,
-                            [
-                                anid,
-                                element.product_id,
-                                element.length,
-                                element.quantity,
-                                element.total_length,
-                                element.module,
-                                element.area,
-                                created_date,
-                                updated_date
-                            ],
-                            (error,results,data)=>{
-                                console.log(error,"err2");
-                                if(error){
-                                    return callBack(error);
-                                }
-                                if(0 == --len){
-                                    return callBack(null,anid + " Annexure Created!!")
-                                }
+                    var promiseArray=[];
+                    if(len !== 0){
+                        prodinfo.forEach(element => {
+                            var promiseFun=new Promise((resolve,reject)=>{
+                                pool.query(
+                                    `insert into annexure_details (annexure_id,product_id,length,quantity,total_length,module,area,created_date,updated_date,status) values(?,?,?,?,?,?,?,?,?,1)`,
+                                    [
+                                        anid,
+                                        element.product_id,
+                                        element.length,
+                                        element.quantity,
+                                        element.total_length,
+                                        element.module,
+                                        element.area,
+                                        created_date,
+                                        updated_date
+                                    ],
+                                    (error,results,data)=>{
+                                        console.log(error,"err2");
+                                        if(error){
+                                            reject(error);
+                                        }
+                                        resolve(results);
+                                    }
+                                )
+                            })
+                            promiseArray.push(promiseFun);
+                        });
+
+                        Promise.all(promiseArray).then((values)=>{
+                            if(acinfo.length != 0){
+                                acinfo.forEach(element => {
+                                    // console.log(element,"element");
+                                    pool.query(
+                                        `INSERT INTO accessories(annexure_id,accessories_name,accessories_quantity,accessories_length,accessories_total_length,accessories_module,accessories_area,created_date,updated_date) values(?,?,?,?,?,?,?,?,?)`,
+                                        [
+                                            anid,
+                                            element.accessories_name,
+                                            element.accessories_quantity,
+                                            element.accessories_length,
+                                            element.accessories_total_length,
+                                            element.accessories_module,
+                                            element.accessories_area,
+                                            created_date,
+                                            updated_date
+                                        ],
+                                        (error,results,fields)=>{
+                                            // console.log(error,"error");
+                                            // console.log(results,"results");
+                                            if(error){
+                                                return callBack(error);
+                                            }
+                        
+                                            if(0 == --acinfo.length){
+                                                return callBack(null,anid);
+                                            }
+                                        }
+                                    )
+                                });
+                            }else{
+                                return callBack(null,anid);
                             }
-                        )
-                    });
+                        })
+                    }else{
+                        if(acinfo.length != 0){
+                            acinfo.forEach(element => {
+                                // console.log(element,"element");
+                                pool.query(
+                                    `INSERT INTO accessories(annexure_id,accessories_name,accessories_quantity,accessories_length,accessories_total_length,accessories_module,accessories_area,created_date,updated_date) values(?,?,?,?,?,?,?,?,?)`,
+                                    [
+                                        anid,
+                                        element.accessories_name,
+                                        element.accessories_quantity,
+                                        element.accessories_length,
+                                        element.accessories_total_length,
+                                        element.accessories_module,
+                                        element.accessories_area,
+                                        created_date,
+                                        updated_date
+                                    ],
+                                    (error,results,fields)=>{
+                                        // console.log(error,"error");
+                                        // console.log(results,"results");
+                                        if(error){
+                                            return callBack(error);
+                                        }
+                    
+                                        if(0 == --acinfo.length){
+                                            return callBack(null,anid);
+                                        }
+                                    }
+                                )
+                            });
+                        }else{
+                            return callBack(null,anid);
+                        }
+                    }
                 }else{
                     return callBack(null,"Annexure not created!!")
                 }
             }
         )
+        }else{
+            return callBack(null,"Annexure not created!!")
+        }
     },
     getAnnexure_details:callBack=>{
         pool.query(
@@ -115,9 +193,76 @@ module.exports={
     //         }
     //     );
     // },
+    // updateAnnexure_details:(body,callBack)=>{
+    //     // console.log(body,"body");
+    //     var cur=new Date().toLocaleString('en-US',{timeZone:'Asia/Calcutta'});
+    //     updated_date=DATE_FORMATTER(cur,"yyyy-mm-dd hh:MM:ss");
+    //     pool.query(
+    //         `update annexure set updated_date=? where annexure_id=?`,
+    //         [
+    //             updated_date,
+    //             body.annexure_id
+    //         ],
+    //         (error,results,data)=>{
+    //             // console.log(error,"error");
+    //             if(error){
+    //                 return callBack(error);
+    //             }
+    //             if(results.affectedRows != 0){
+    //                 var product=results.affectedRows;
+    //                 pool.query(
+    //                     `delete from annexure_details, accessories_details where annexure_id=?`,
+    //                     [
+    //                      body.annexure_id   
+    //                     ],
+    //                     (error,results,fields)=>{
+    //                         if(error){
+    //                             return callBack(error);
+    //                         }
+    //                 if(results.affectedRows != 0){
+                        
+    //                     console.log(body,"body");
+    //                     body.productinfo.forEach(element => {
+    //                     pool.query(
+    //                         `INSERT INTO annexure_details(annexure_id,product_id,length,quantity,total_length,module,area,created_date,updated_date,status) VALUES (?,?,?,?,?,?,?,?,?,1)`,
+    //                         [
+    //                          body.annexure_id,
+    //                          element.product_id,
+    //                          element.length,
+    //                          element.quantity,
+    //                          element.total_length,
+    //                          element.module,
+    //                          element.area,
+    //                          created_date,
+    //                          updated_date
+    //                         ],
+    //                         (error,results,fields)=>{
+    //                             // console.log(error,"error");
+    //                             if(error){
+    //                                 return callBack(error);
+    //                             }
+    //                             if(0 == --body.productinfo.length){
+    //                                 return callBack(null,product);
+    //                             }
+    //                         }
+    //                     )
+    //                 });
+    //             }else{
+    //             return callBack(null,results);
+    //             }
+    //         }
+    //     )
+    //     }else{
+    //         return callBack(null,results)
+    //     }
+    //             //return callBack(null,results);
+    //     }
+    //     )
+        
+    // },
     updateAnnexure_details:(body,callBack)=>{
+        // console.log(body,"body");
         var cur=new Date().toLocaleString('en-US',{timeZone:'Asia/Calcutta'});
-        created_date=DATE_FORMATTER(cur,"yyyy-mm-dd hh:MM:ss");
         updated_date=DATE_FORMATTER(cur,"yyyy-mm-dd hh:MM:ss");
         pool.query(
             `update annexure set updated_date=? where annexure_id=?`,
@@ -131,9 +276,36 @@ module.exports={
                     return callBack(error);
                 }
                 if(results.affectedRows != 0){
-                    var product=results.affectedRows;
+                    new Promise((resolve,reject)=>{
+                        pool.query(
+                            `delete from annexure_details where annexure_id=?`,
+                            [
+                             body.annexure_id   
+                            ],
+                            (error,results,fields)=>{
+                                if(error){
+                                    reject(error);
+                                }
+                                resolve(results);
+                            }).then((values)=>{
+                                new Promise((resolve,reject)=>{
+                                    pool.query(
+                                        `delete from accessories where annexure_id=?`,
+                                        [body.annexure_id],
+                                        (error,respp,fields)=>{
+                                            if(error){
+                                                reject(error);
+                                                return callBack(error);
+                                            }
+                                            resolve(respp);
+                                            return callBack(null,respp);
+                                        }
+                                    )
+                                })
+                            })                        
+                    })
                     pool.query(
-                        `delete from annexure_details where annexure_id=?`,
+                        `delete from annexure_details, accessories_details where annexure_id=?`,
                         [
                          body.annexure_id   
                         ],
@@ -142,7 +314,8 @@ module.exports={
                                 return callBack(error);
                             }
                     if(results.affectedRows != 0){
-                        // console.log(body,"body");
+                        
+                        console.log(body,"body");
                         body.productinfo.forEach(element => {
                         pool.query(
                             `INSERT INTO annexure_details(annexure_id,product_id,length,quantity,total_length,module,area,created_date,updated_date,status) VALUES (?,?,?,?,?,?,?,?,?,1)`,
@@ -181,18 +354,47 @@ module.exports={
         )
         
     },
-    deleteAnnexure_details:(id,callBack)=>{
+    delete_annexure_and_details:(id,callBack)=>{
         pool.query(
-            `delete from annexure_details where annexure_details_id=?`,
-            [id],
-            (error,results,fields)=>{
+            `delete from annexure where annexure_id=?`,
+            [
+                id
+            ],
+            (error,results,data)=>{
+                // console.log(error,"error");
                 if(error){
                     return callBack(error);
                 }
-                return callBack(null,results);
-            }
-        );
-    },
+                if(results.affectedRows != 0){
+                    new Promise((resolve,reject)=>{
+                        pool.query(
+                            `delete from annexure_details where annexure_id=?`,
+                            [
+                             id   
+                            ],
+                            (error,results,fields)=>{
+                                if(error){
+                                    reject(error);
+                                }
+                                resolve(results);
+                            })
+                        }).then((values)=>{
+                                    pool.query(
+                                        `delete from accessories where annexure_id=?`,
+                                        [id],
+                                        (error,respp,fields)=>{
+                                            console.log(error);
+
+                                            if(error){
+                                                return callBack(error);
+                                            }
+                                            return callBack(null,respp);
+                                        }
+                                    )
+                                })                    
+    }
+}
+)},
     getAnnexure_detailsby_annexure_id:(id,callBack)=>{
         pool.query(
             `select ad.annexure_details_id,p.product_name,ad.product_id,ad.length,quantity,ad.total_length,ad.module,ad.area,ad.created_date,ad.updated_date,ad.status from annexure_details ad join product p on ad.product_id=p.product_id where ad.annexure_id=?`,
@@ -214,34 +416,34 @@ module.exports={
     },
 
 
-    delete_annexure_and_details:(id,callBack)=>{
-        // console.log(id,"id");
-        pool.query(
-            'delete from annexure where annexure_id=?',
-            [id],
-            (error,results,data1)=>{
-                // console.log(error,"err")
-                if(error){
-                    return callBack(error);
-                }
-                if(results.affectedRows != 0){
-                    // console.log("in");
-                    pool.query(
-                        `delete from annexure_details where annexure_id=?`,
-                        [id],
-                        (error,results,fields)=>{
-                            if(error)
-                            {
-                            return callBack(error);
-                            }
-                            return callBack(null,results);
-                        }
-                    )
-                }
+    // delete_annexure_and_details:(id,callBack)=>{
+    //     // console.log(id,"id");
+    //     pool.query(
+    //         'delete from annexure where annexure_id=?',
+    //         [id],
+    //         (error,results,data1)=>{
+    //             // console.log(error,"err")
+    //             if(error){
+    //                 return callBack(error);
+    //             }
+    //             if(results.affectedRows != 0){
+    //                 // console.log("in");
+    //                 pool.query(
+    //                     `delete from annexure_details where annexure_id=?`,
+    //                     [id],
+    //                     (error,results,fields)=>{
+    //                         if(error)
+    //                         {
+    //                         return callBack(error);
+    //                         }
+    //                         return callBack(null,results);
+    //                     }
+    //                 )
+    //             }
 
-            }
-        )
-    },
+    //         }
+    //     )
+    // },
     get_annexure_details_json:(id,callBack)=>{
         var resultRow=[];
         // console.log(id,"id");
